[共通] refactor:インタフェースをtypeに変更等(priorityUtils.ts)
インタフェースをtypeに変更
関数名変更
呼び出し元含め、構造を整理
